refactor: extract workflow footer status
