+++ ui/application/__init__.py
@@ -115,6 +115,11 @@ from .local_first_progress_facts import (
     current_local_first_last_sync_date,
     runtime_state_with_local_first_output_path,
 )
+from .workflow_footer_status import (
+    WorkflowFooterFacts,
+    build_workflow_footer_facts_from_progress_facts,
+    build_workflow_footer_status,
+)
 from .workflow_progress import (
     build_startup_workflow_progress_facts,
     build_workflow_progress_from_facts,
@@ -258,6 +263,7 @@ __all__ = [
     "VisualWorkflowActionInputs",
     "VisualWorkflowSettingsSnapshot",
     "WizardProgressFacts",
+    "WorkflowFooterFacts",
     "WorkflowProgressFacts",
     "WizardSettingsSnapshot",
     "build_advanced_fetch_visibility_update",
@@ -290,6 +296,8 @@ __all__ = [
     "build_startup_workflow_progress_facts",
     "current_local_first_last_sync_date",
     "build_wizard_filter_description",
+    "build_workflow_footer_facts_from_progress_facts",
+    "build_workflow_footer_status",
     "build_workflow_progress_facts_from_runtime_state",
     "build_workflow_progress_from_facts",
     "build_workflow_progress_from_facts_and_settings",
